requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
pre-commit==3.3.3
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
python-dotenv==1.0.0
django-redis==5.4.0
factory-boy==3.3.0
